requirements.txt
streamlit==1.28.0
torch==2.1.0
torchvision==0.18.1
Pillow==10.0.0
